+++ integration_tests/suite/helpers/database.py
@@ -1,4 +1,4 @@
-# Copyright 2015-2023 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2015-2024 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from contextlib import contextmanager
@@ -430,6 +430,54 @@ class DatabaseQueries:
         )
 
         return filter_member_id
+
+    def insert_parking_lot(
+        self,
+        number='700',
+        context='default',
+        slots_start='701',
+        slots_end='750',
+        tenant_uuid=None,
+    ):
+        query = text(
+            """
+        INSERT INTO parking_lot (slots_start, slots_end, tenant_uuid)
+        VALUES (:slots_start, :slots_end, :tenant_uuid)
+        RETURNING id
+        """
+        )
+
+        parking_lot_id = self.connection.execute(
+            query,
+            slots_start=slots_start,
+            slots_end=slots_end,
+            tenant_uuid=tenant_uuid,
+        ).scalar()
+        self.insert_extension(number, context, 'parking', parking_lot_id)
+        return parking_lot_id
+
+    def delete_parking_lot(self, parking_lot_id, extension_id=None):
+        query = text(
+            "DELETE FROM func_key_dest_parking WHERE parking_lot_id = :parking_lot_id"
+        )
+        self.connection.execute(query, parking_lot_id=parking_lot_id)
+
+        query = text(
+            """
+            DELETE FROM func_key_dest_park_position
+            WHERE parking_lot_id = :parking_lot_id
+            """
+        )
+        self.connection.execute(query, parking_lot_id=parking_lot_id)
+
+        query = text("DELETE FROM parking_lot WHERE id = :parking_lot_id")
+        self.connection.execute(query, parking_lot_id=parking_lot_id)
+        if extension_id:
+            self.delete_extension(extension_id)
+
+    def delete_extension(self, exten):
+        query = text("DELETE FROM extensions WHERE exten = :exten")
+        self.connection.execute(query, exten=exten)
 
     def delete_context(self, name):
         query = text("DELETE FROM context WHERE name = :name")