+++ wazo_confd/plugins/func_key/resource.py
@@ -15,6 +15,8 @@ from xivo_dao.alchemy.func_key_dest_forward import FuncKeyDestForward
 from xivo_dao.alchemy.func_key_dest_group import FuncKeyDestGroup
 from xivo_dao.alchemy.func_key_dest_group_member import FuncKeyDestGroupMember
 from xivo_dao.alchemy.func_key_dest_paging import FuncKeyDestPaging
+from xivo_dao.alchemy.func_key_dest_park_position import FuncKeyDestParkPosition
+from xivo_dao.alchemy.func_key_dest_parking import FuncKeyDestParking
 from xivo_dao.alchemy.func_key_dest_queue import FuncKeyDestQueue
 from xivo_dao.alchemy.func_key_dest_service import FuncKeyDestService
 from xivo_dao.alchemy.func_key_dest_user import FuncKeyDestUser
@@ -42,8 +44,8 @@ models_destination = {
     'groupmember': FuncKeyDestGroupMember,
     'onlinerec': FuncKeyDestOnlineRecording,
     'paging': FuncKeyDestPaging,
-    # 'park_position': FuncKeyDestParkPosition,
-    # 'parking': FuncKeyDestParking,
+    'park_position': FuncKeyDestParkPosition,
+    'parking': FuncKeyDestParking,
     'queue': FuncKeyDestQueue,
     'service': FuncKeyDestService,
     'transfer': FuncKeyDestTransfer,